+++ examples/create-invoice.ts
@@ -1,57 +1,62 @@
-const CrypoBotAPI = require('crypto-bot-api');
+import CrypoBotAPI, { type CreateInvoiceOptions } from 'crypto-bot-api';
 
 const client = new CrypoBotAPI('1234:AAA...AAA');
 
 // Only required parameters
-client.createInvoice({
+const minimalOptions: CreateInvoiceOptions = {
   amount: 9.5,
   asset: 'TON'
-}).then(invoice => {
+};
+client.createInvoice(minimalOptions).then(invoice => {
   // Send invoice.payUrl to user
   console.log(invoice);
 });
 
 // With pay in any support crypto currency using exchange rate passed fiat currency
-client.createInvoice({
+const fiatOptions: CreateInvoiceOptions = {
   amount: 99,
   currencyType: CrypoBotAPI.CurrencyType.Fiat,
   fiat: 'USD'
-}).then(invoice => {
+}
+client.createInvoice(fiatOptions).then(invoice => {
   // Send invoice.payUrl to user
   console.log(invoice);
 });
 
 // With pay in passed crypto currencies using exchange rate passed fiat currency
-client.createInvoice({
+const fiatLimitedOptions: CreateInvoiceOptions = {
   amount: 99,
   currencyType: CrypoBotAPI.CurrencyType.Fiat,
   fiat: 'USD',
   acceptedAssets: ['TON', 'BTC']
-}).then(invoice => {
+};
+client.createInvoice(fiatLimitedOptions).then(invoice => {
   // Send invoice.payUrl to user
   console.log(invoice);
 });
 
 // With description, payload and anonymous payment disabled
-client.createInvoice({
+const custom1Options: CreateInvoiceOptions = {
   amount: 9.5,
   asset: 'TON',
   description: 'Pay order in SuperShop',
   isAllowAnonymous: false,
   payload: 'order12345'
-}).then(invoice => {
+};
+client.createInvoice(custom1Options).then(invoice => {
   // Send invoice.payUrl to user
   console.log(invoice);
 });
 
 // With pay button and payment comment disabled
-client.createInvoice({
+const custom2Options: CreateInvoiceOptions = {
   amount: 9.5,
   asset: 'TON',
   paidBtnName: 'viewItem',
   paidBtnUrl: 'https://supershop.com/order12345',
   isAllowComments: false,
-}).then(invoice => {
+};
+client.createInvoice(custom2Options).then(invoice => {
   // Send invoice.payUrl to user
   console.log(invoice);
 });