+++ examples/exchange-rate-ticker.ts
@@ -1,4 +1,4 @@
-const CrypoBotAPI = require('crypto-bot-api');
+import CrypoBotAPI from 'crypto-bot-api';
 
 const client = new CrypoBotAPI('1234:AAA...AAA');