+++ examples/get-invoices-paginate.js
@@ -3,9 +3,4 @@ const CrypoBotAPI = require('crypto-bot-api');
 const client = new CrypoBotAPI('1234:AAA...AAA');
 client.setPageSize(20);
 
-client.getInvoicesPaginate({ asset: 'BTC', page: 2 })
-  .then(invoices => {
-    console.log(invoices.page)
-    console.log(invoices.pagesCount)
-    console.log(invoices.items)
-  });
+client.getInvoicesPaginate({ asset: 'BTC', page: 2 }).then(invoices => console.log(invoices));