+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - Add new support `MEMHASH` currency code to types
+- Fix `paid` event arguments optional status
 - Update dependencies, fix build config
 
 ## 0.3.4 - 2025-01-30