+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Add new support `MEMHASH` currency code to types
 - Fix `paid` event arguments optional status
+- Fix `http` create server field required status
 - Update dependencies, fix build config
 
 ## 0.3.4 - 2025-01-30